Major fix
not all btaf entries have a "bits" uint, and are 3 uints long, not 4. this caused the reader to become misaligned with the btaf entries. Now GARCUnpack checks the first uint for if it's a "bits" or not. Also it now extracts a file for every btaf entry, not every otaf entry. this is because some garc files have more btafs than otafs. Also corrected "bmif" being misspelled as "gmif"